Read style props through `StyleInfo` in  `AdjustCssLengthProperties` (#6632)
## Problem
The padding strategy relies on the `AdjustCssLengthProperties` command
to update elements, both for reading and writing styles. In order for
that to work in Tailwind projects, where elements don't have an inline
style prop, `AdjustCssLengthProperties` needs to read element styles
from the "right" property, otherwise the command wouldn't work as
intended.

## Fix
Use the `StyleInfo` system from to read the element style info.

### Details
- Refactored `AdjustCssLengthProperties` to read elements styles through
`styleInfo`
- Created a bespoke prop, `LengthProperty`, to track which props are
addressed by `LengthPropertyToAdjust`. This way it's easy to tell which
props need to be supported by `StyleInfo` to have
`AdjustCssLengthProperties` working well
- Extended `StyleInfo` to support the props read by
`AdjustCssLengthProperties`
- Updated the style plugins to support the new `StyleInfo` props

### Manual Tests:
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Play mode