feat(loading): merge github fetch with the loading bar (#6630)
This PR adds the loading stages to the loading bar screen and progress
bar.

Stages for GH import:
1. Loading Editor...
2. Fetching <repo name>
3. Parsing files
4. Fetching dependencies

Stages for project load (we reverse the order in our project load):
1. Fetching dependencies
2. Parsing files

If everything is validated correctly - it just shows the editor with the
project ready to be worked on. If not - it shows an error modal (its UI
was slightly changed in #6677 to better reflect errors).

It also adds support for Dark mode.

**Note:**
- The new logic is almost entirely in
[editor/src/components/editor/loading-screen.tsx](https://github.com/concrete-utopia/utopia/pull/6630/files#diff-5aa4d7e811e0bfb07a60910e517fe3a01cd270c9b4fd05132e513d939d69f769)
- There is a jump in the progress bar since these are actually two DOM
elements, one that is rendered before the React is loaded and a React
component that replaces it after the React code is loaded. This jump
will be fixed in a subsequent PR.
- Currently the progress bar is "static" (advances by time and not
stages). This will be changed in (the same) subsequent PR.

<video
src="https://github.com/user-attachments/assets/a6d52f3d-b27b-4110-be1e-a7f213f563ab"></video>

**Manual Tests:**
I hereby swear that:

- [X] I opened a hydrogen project and it loaded
- [X] I could navigate to various routes in Play mode