feat(editor): add responsive utils (#6716)
This prep PR adds the responsive utilities we'll use in our responsive
PR, along with types and tests.
The important utilities (that are exposed outside) are:
1. `extractScreenSizeFromCss(css: string)` - which recieves a CSS string
of the media query (i.e `@media (min-width: 100px)`) and turns it into a
`ScreenSize` object (in our example - `{ min: {value: 100, unit: 'px'}
}`). We also support media ranges (`@media (20px < width < 50em)`). This
function is covered in tests.
It uses `mediaQueryToScreenSize(mediaQuery: MediaQuery)` internally,
which takes a `MediaQuery` object (after it was parsed from the CSS
string), and does the "heavy lifting" of converting it to a `ScreenSize`
representation. This inner function is also covered in tests.

2. `selectValueByBreakpoint` - this function recieves a list of possible
variants (values with corresponding `ScreenSize`s which they apply in),
and the current Scene size. It infers the most matching variant
according to the Scene size (can also be the default variant if none are
matching). This function is also fully covered in tests.

No functionality is added to the app itself, this is just a prep PR for
the functionality in the subsequent PR.

**Manual Tests:**
I hereby swear that:

- [X] I opened a hydrogen project and it loaded
- [X] I could navigate to various routes in Play mode