feat(editor): pass scene size to style plugins (#6715)
This prep PR adds the ability to pass the Scene size to the Tailwind
plugin/
This was extracted from the spike since it added small changes in many
files.

**Details:**

- We need the Scene size in
[tailwind-style-plugin.ts](https://github.com/concrete-utopia/utopia/pull/6715/files#diff-ad30d4bf205861b9efac9bb4bdb2bb68707cace215f64915244c46dbfab0607dR23-R25)
to be able (in the next PR) to calculate the matching breakpoint.
- We get the Scene size in `getContainingSceneWidth` in
[responsive-utils.ts](https://github.com/concrete-utopia/utopia/pull/6715/files#diff-4f8de694cb6cb8852ff046766f5015a01f9567d4c3b7472bbac36358b52932aaR7-R16)
and we pass it to the Tailwind plugin.
- The rest of the changes are just passing down this data.

This PR doesn't change any functionality, but is merely a prep PR for
the next, functional one.

**Manual Tests:**
I hereby swear that:

- [X] I opened a hydrogen project and it loaded
- [X] I could navigate to various routes in Play mode